aurora_robot/requirements.txt
slackclient
certifi
slackeventsapi
Flask
pymongo